+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * refactor: Mover more logic to `TransformerRegistry`.
 * refactor: Move `MainTransformer` to its own namespace.
 * refactor: Refactor exception.
+* feat: Add attribute matching.
 
 ## 0.5.4